refactor(app): move trivia to its own component
