+++ src/views/Trivia.svelte
@@ -1,0 +1,69 @@
+<script>
+  import { store, questions, currentIndex } from '../store/store';
+  import triviaApi from '../api/trivia';
+  import TriviaCard from '../components/TriviaCard.svelte';
+  import TriviaQuestion from '../components/TriviaQuestion.svelte';
+
+  const AMOUNT_QUESTIONS = 10;
+
+  let difficulty = null;
+
+  const startGame = async (diff) => {
+    await fetchNewQuestions();
+    $store.isPlaying = true;
+  };
+
+  const fetchNewQuestions = async (diff) => {
+    const res = await triviaApi.getQuestions({
+      amount: AMOUNT_QUESTIONS,
+      difficulty: difficulty,
+      type: 'multiple',
+    });
+    $questions = res.data.results;
+    console.log($questions[0]);
+  };
+
+  let maxCorrectAnswers = 0;
+
+  function updateMaxCorrectAnswers(newValue) {
+    if (maxCorrectAnswers < newValue) {
+      $store.maxCorrectAnswers = newValue;
+    }
+  }
+
+  $: updateMaxCorrectAnswers($store.correctAnswersCount);
+</script>
+
+<main>
+  <div class="flex flex-col justify-center text-center">
+    <div>
+      <h1 class="mb-10 text-3xl font-bold underline">Trivia</h1>
+
+      {#if $store.isPlaying}
+        <div>
+          <p class="text-xl">
+            Respuestas Correctas
+            <span class="font-bold">{$store.correctAnswersCount} / {AMOUNT_QUESTIONS} </span>
+          </p>
+          <TriviaQuestion />
+          <p class="text-2xl">
+            {$currentIndex < 10 ? $currentIndex + 1 : AMOUNT_QUESTIONS} / {AMOUNT_QUESTIONS}
+          </p>
+        </div>
+      {:else}
+        <TriviaCard>
+          <div>
+            <p>Elige la dificultad</p>
+            <select bind:value={difficulty} class="mb-10">
+              <option disabled value="">Dificultad</option>
+              <option value="easy">Fácil</option>
+              <option value="medium">Mediana</option>
+              <option value="hard">Difícil</option>
+            </select>
+          </div>
+          <button on:click={startGame} class="p-3 bg-green-400 rounded-lg ">Empezar juego</button>
+        </TriviaCard>
+      {/if}
+    </div>
+  </div>
+</main>
